// src/interfaces/CategoriesInterfaces.ts
export interface CategoriesRes {
  results: number;
  metadata: Metadata;
  data: Category[];
}

export interface Metadata {
  currentPage: number;
  numberOfPages: number;
  limit: number;
}

export interface Category {
  _id: string;
  name: string;
  slug: string;
  image: string;
  createdAt: string;
  updatedAt: string;
}